requirements.txt
torch==2.0.1
torchvision==0.15.2
numpy==1.23.5
matplotlib==3.7.1
pandas==2.0.0
tqdm==4.65.0
pyyaml==6.0